feat: working prototype
