+++ cmd/root.go
@@ -1,47 +1,41 @@
 package cmd
 
 import (
-	"errors"
 	"fmt"
 	"os"
 
-	"github.com/sethpollack/ecr-cleaner/ecr"
+	"github.com/ltvco/ecr-cleaner/clean"
 	"github.com/spf13/cobra"
 )
 
 var (
-	opts ecr.Opts
-
-	runExample = `
-		ecr-cleaner \
-			-r "(^[0-9a-f]+$|-[0-9a-f]+$)" \
-			-n foobar
-	`
-	rootCmd = &cobra.Command{
-		Use:   "ecr-cleaner",
-		Short: "Clean untagged images from ECR",
-		Long:  runExample,
-		PreRunE: func(cmd *cobra.Command, args []string) error {
-			if !opts.RunAll && opts.RepoName == "" {
-				return errors.New("You must specify either --repo-name or --all.")
-			} else {
-				return nil
-			}
-		},
-		RunE: func(cmd *cobra.Command, args []string) error {
-			return ecr.CleanRepos(opts)
-		},
-	}
+	untaggedOnly bool
+	dryRun       bool
+	keepLast     int
+	profileName  string
+	region       string
 )
 
 func init() {
-	rootCmd.Flags().StringVarP(&opts.TagRegex, "tag-regex", "r", "(^[0-9a-f]+$|-[0-9a-f]+$)", "Image tag regex")
-	rootCmd.Flags().StringVarP(&opts.RepoName, "repo-name", "n", "", "Name of repo to clean.")
-	rootCmd.Flags().BoolVarP(&opts.RunAll, "all", "a", false, "Clean all repos.")
-	rootCmd.Flags().BoolVarP(&opts.DryRun, "dry-run", "d", true, "dry run to see tags that will be deleted.")
+
+	rootCmd.PersistentFlags().BoolVarP(&untaggedOnly, "untagged-only", "u", true, "Only delete untagged images (default: true)")
+	rootCmd.PersistentFlags().IntVarP(&keepLast, "keep-latest-count", "k", 10, "Keep the latest X images (default: 10)")
+	rootCmd.PersistentFlags().StringVarP(&profileName, "profile-name", "p", "NA", "The AWS profile to use (default: '')")
+	rootCmd.PersistentFlags().StringVarP(&region, "region", "r", "us-east-1", "AWS region (default: us-east-1)")
+	rootCmd.PersistentFlags().BoolVarP(&dryRun, "dry-run", "d", true, "Dry run of the clean action (default: true)")
+
+}
+
+var rootCmd = &cobra.Command{
+	Use:   "ecr-cleaner",
+	Short: "clean amazon elastic container registries",
+	Run: func(cmd *cobra.Command, args []string) {
+		clean.CleanRepos(untaggedOnly, keepLast, profileName, region, dryRun)
+	},
 }
 
 func Execute() {
+
 	if err := rootCmd.Execute(); err != nil {
 		fmt.Println(err)
 		os.Exit(1)
